Fusion and Step files for side fan shroud idea
Working on a side blower fan solution using all stock M3D parts.  This file does not yet have an optimized blower outlet, it is for testing machines limits and clearances.  Once the community is happy with sizing we can create new blower outlets optimized for the K'Tana or Compound setups.